Split inventory service and report modules
